backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.0.1
pytesseract==0.3.10
numpy==1.24.3
pymupdf==1.23.8
python-multipart==0.0.6
pydantic==2.5.0
google-generativeai==0.7.1
google-generativeai==0.7.1
